+++ src/main/java/de/dm/intellij/liferay/language/poshi/annotation/PoshiValidatorRunnerParams.java
@@ -1,0 +1,11 @@
+package de.dm.intellij.liferay.language.poshi.annotation;
+
+import com.intellij.openapi.project.Project;
+
+import java.io.File;
+
+public class PoshiValidatorRunnerParams {
+
+	public Project project;
+	public File workingDirectory;
+}